+++ docs/src/routes/(docs)/[...slug]/+page.svelte
@@ -32,6 +32,7 @@
     import RadioGroupDemo from "$lib/components/demos/RadioGroupDemo.svelte";
     import SelectDemo from "$lib/components/demos/SelectDemo.svelte";
     import SwitchDemo from "$lib/components/demos/SwitchDemo.svelte";
+    import TableDemo from "$lib/components/demos/TableDemo.svelte";
     import TabsDemo from "$lib/components/demos/TabsDemo.svelte";
     import TextareaDemo from "$lib/components/demos/TextareaDemo.svelte";
     import ToggleDemo from "$lib/components/demos/ToggleDemo.svelte";
@@ -78,6 +79,7 @@
         RadioGroupDemo,
         SelectDemo,
         SwitchDemo,
+        TableDemo,
         TabsDemo,
         TextareaDemo,
         ToggleDemo,